Falta comments (ambos) y postLikes

+++ tp/seed/social-network/src/main/java/io/redbee/socialnetwork/users/controller/UserController.java
@@ -1,6 +1,6 @@
 package io.redbee.socialnetwork.users.controller;
 
-import io.redbee.socialnetwork.configuration.hateoas.PaginatedResultsRetrievedEvent;
+import io.redbee.socialnetwork.shared.controller.SecuredController;
 import io.redbee.socialnetwork.users.mapper.UserToResponseMapper;
 import io.redbee.socialnetwork.users.model.User;
 import io.redbee.socialnetwork.users.model.UserCreationRequest;
@@ -10,35 +10,42 @@ import io.redbee.socialnetwork.users.service.UserDeleteService;
 import io.redbee.socialnetwork.users.service.UserSearchService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.util.List;
-import java.util.stream.Collectors;
+
+import static org.springframework.http.HttpStatus.CREATED;
 
 @RestController
 @RequestMapping("/users")
-public class UserController {
+@PreAuthorize("isAuthenticated()")
+public class UserController extends SecuredController {
 
     private final UserCreationService creationService;
     private final UserSearchService searchService;
     private final UserDeleteService deleteService;
     private final UserToResponseMapper responseMapper;
-    private final ApplicationEventPublisher eventPublisher;
 
     private static final Logger LOGGER = LoggerFactory.getLogger(UserController.class);
 
-    public UserController(UserCreationService creationService, UserSearchService searchService, UserDeleteService deleteService, UserToResponseMapper responseMapper, ApplicationEventPublisher eventPublisher) {
+    public UserController(
+            UserSearchService searchService,
+            UserCreationService creationService,
+            UserDeleteService deleteService,
+            UserToResponseMapper responseMapper
+    ) {
+        super(searchService);
         this.creationService = creationService;
         this.searchService = searchService;
         this.deleteService = deleteService;
         this.responseMapper = responseMapper;
-        this.eventPublisher = eventPublisher;
     }
 
     @GetMapping()
@@ -48,44 +55,47 @@ public class UserController {
             HttpServletResponse response
     ) {
         LOGGER.info("get: searching users");
-        Page<User> responsePage = searchService.get(pageable);
 
-        eventPublisher.publishEvent(new PaginatedResultsRetrievedEvent<User>(
-                User.class, uriBuilder, response, pageable.getPageNumber(), responsePage.getTotalPages(), pageable.getPageSize()
-        ));
+        Page<User> responsePage = searchService.getPage(pageable);
 
-        LOGGER.info("get: returning users");
-        return responsePage.getContent()
-                .stream()
+        LOGGER.info("get: users found");
+        return responsePage
                 .map(this::mapResponse)
-                .collect(Collectors.toList());
+                .getContent();
     }
 
     @GetMapping("/{id}")
     public UserResponse getById(@PathVariable Integer id) {
         LOGGER.info("getById: searching for user {}", id);
-        UserResponse response = mapResponse(searchService.get(id));
+        UserResponse response = mapResponse(searchService.getBy(id));
         LOGGER.info("getById: user found {}", response);
         return response;
     }
 
     @PostMapping()
-    public UserResponse create(@RequestBody @Valid UserCreationRequest request) {
-        LOGGER.info("create: creating user {}", request.getMail());
-        UserResponse response = mapResponse(creationService.create(request.getMail(), request.getPassword()));
-        LOGGER.info("create: created user {}", response);
+    @ResponseStatus(CREATED)
+    @PreAuthorize("permitAll()")
+    public UserResponse create(@RequestBody @Valid UserCreationRequest user) {
+        LOGGER.info("create: creating user {}", user.getMail());
+        UserResponse response = mapResponse(
+                creationService.create(user.getMail(), user.getPassword())
+        );
+        LOGGER.info("create: user created {}", response);
         return response;
     }
 
     @DeleteMapping("/{id}")
     public UserResponse delete(@PathVariable Integer id) {
+        validateUserIsOwner(id);
         LOGGER.info("delete: deleting user {}", id);
+        SecurityContextHolder.getContext().getAuthentication().getDetails();
         UserResponse response = mapResponse(deleteService.delete(id));
-        LOGGER.info("delete: deleted user: {}", response);
+        LOGGER.info("delete: user deleted: {}", response);
         return response;
     }
 
     private UserResponse mapResponse(User user) {
         return responseMapper.map(user);
     }
+
 }
